(refactor) Moves trajectory utility functions in trajectory.h

+++ src/kompass_cpp/kompass_cpp/include/utils/cost_evaluator.h
@@ -185,8 +185,8 @@ private:
                                      // respect to the robot
 
 #ifdef GPU
-  int numTrajectories_;
-  int numPointsPerTrajectory_;
+  size_t numTrajectories_;
+  size_t numPointsPerTrajectory_;
   double *m_devicePtrPathsX;
   double *m_devicePtrPathsY;
   double *m_devicePtrVelocitiesVx;
